+++ sqlalchemy_exasol/warnings.py
@@ -1,6 +1,3 @@
-from warnings import UserWarning, DeprecationWarning
-
-
 class SqlaExasolWarning(UserWarning):
     """Base class for all warnings emited by sqlalchemy_exasol."""